Resolved Form types should be in Form/Type folders
