+++ src/AppBundle/Form/Type/EditQuestionType.php
@@ -1,32 +1,28 @@
 <?php
 
-namespace AppBundle\Form;
+namespace AppBundle\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use AppBundle\Repository\QuestionRepository;
+use AppBundle\Repository\PoolQuestionRepository;
 
 use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
+use Symfony\Component\Form\Extension\Core\Type\FileType;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 
 
 
-class EditResponseType extends AbstractType
+class EditQuestionType extends AbstractType
 {
     public function buildform(FormBuilderInterface $builder, array $option)
     {
         $builder->add('title', TextType::class)
-                ->add('isFair', CheckboxType::class, array(
-                    'label' =>  'Bonne réponse ',
-                    'required'  =>  false
-                ))
-                ->add('question', EntityType::class, array(
-                        'class' => 'AppBundle:Question',
-                        'query_builder' => function (QuestionRepository $qr) {
-                            return $qr->createQueryBuilder('pr');
+                ->add('poolQuestion', EntityType::class, array(
+                        'class' => 'AppBundle:PoolQuestion',
+                        'query_builder' => function (PoolQuestionRepository $pqr) {
+                            return $pqr->createQueryBuilder('pq');
                         },
                         'choice_label' => 'title',
                 )) 
@@ -37,7 +33,7 @@ class EditResponseType extends AbstractType
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults(array(
-            'data_class' => 'AppBundle\Entity\Response'
+            'data_class' => 'AppBundle\Entity\Question'
         ));
     }
 }
