+++ src/AppBundle/Form/Type/PoolQuestionType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppBundle\Form;
+namespace AppBundle\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
@@ -8,22 +8,24 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
 
 use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\Extension\Core\Type\PercentType;
 
 
 
 
-class PoolVideoType extends AbstractType
+class PoolQuestionType extends AbstractType
 {
     public function buildform(FormBuilderInterface $builder, array $option)
     {
         $builder->add('title', TextType::class)
+                ->add('rate', PercentType::class, array('type' => 'integer'))
         ;        
     }
 
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults(array(
-            'data_class' => 'AppBundle\Entity\PoolVideo'
+            'data_class' => 'AppBundle\Entity\PoolQuestion'
         ));
     }
 }
