+++ src/AppBundle/Form/Type/QuestionType.php
@@ -1,32 +1,30 @@
 <?php
 
-namespace AppBundle\Form;
+namespace AppBundle\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use AppBundle\Repository\PoolVideoRepository;
+use AppBundle\Repository\PoolQuestionRepository;
 
 use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\Extension\Core\Type\FileType;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 
 
 
 
-class VideoType extends AbstractType
+class QuestionType extends AbstractType
 {
     public function buildform(FormBuilderInterface $builder, array $option)
     {
         $builder->add('title', TextType::class)
                 ->add('file', FileType::class)
-                ->add('description', TextareaType::class)
-                ->add('poolVideo', EntityType::class, array(
-                        'class' => 'AppBundle:PoolVideo',
-                        'query_builder' => function (PoolVideoRepository $pvr) {
-                            return $pvr->createQueryBuilder('pv');
+                ->add('poolQuestion', EntityType::class, array(
+                        'class' => 'AppBundle:PoolQuestion',
+                        'query_builder' => function (PoolQuestionRepository $pqr) {
+                            return $pqr->createQueryBuilder('pq');
                         },
                         'choice_label' => 'title',
                 ) 
@@ -38,7 +36,7 @@ class VideoType extends AbstractType
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults(array(
-            'data_class' => 'AppBundle\Entity\Video'
+            'data_class' => 'AppBundle\Entity\Question'
         ));
     }
 }
